output
Enter the number of processes
5
Enter the priority of the process
4
Enter the burst time of the process
6
Enter the priority of the process
7
Enter the burst time of the process
8
Enter the priority of the process
9
Enter the burst time of the process
3
Enter the priority of the process
4
Enter the burst time of the process
5
Enter the priority of the process
6
Enter the burst time of the process
7
Queue 1 :       7->9->NULL
Queue 2 :       4->4->6->NULL
Queue 3 :       NULL
Queue 1 in hand
Executing queue 1 and 1 process for a unit time. Process has priority of 7
Executing queue 1 and 1 process for a unit time. Process has priority of 7
Executing queue 1 and 1 process for a unit time. Process has priority of 7
Executing queue 1 and 1 process for a unit time. Process has priority of 7
Executing queue 1 and 2 process for a unit time. Process has priority of 9
Executing queue 1 and 2 process for a unit time. Process has priority of 9
Executing queue 1 and 2 process for a unit time. Process has priority of 9
Broke from queue 1
Queue 2 in hand
Executing queue 2 and 1 process for a unit time. Process has priority of 6
Executing queue 2 and 1 process for a unit time. Process has priority of 6
Executing queue 2 and 1 process for a unit time. Process has priority of 6
Broke from queue 2
Queue 3 completed
Queue 2 in hand
Executing queue 2 and 1 process for a unit time. Process has priority of 6
Executing queue 2 and 1 process for a unit time. Process has priority of 6
Executing queue 2 and 1 process for a unit time. Process has priority of 6
Executing queue 2 and 1 process for a unit time. Process has priority of 6
Executing queue 2 and 2 process for a unit time. Process has priority of 4
Executing queue 2 and 2 process for a unit time. Process has priority of 4
Executing queue 2 and 2 process for a unit time. Process has priority of 4
Executing queue 2 and 2 process for a unit time. Process has priority of 4
Executing queue 2 and 2 process for a unit time. Process has priority of 4
Executing queue 2 and 2 process for a unit time. Process has priority of 4
Broke from queue 2
Queue 3 completed
Queue 2 in hand
Executing queue 2 and 3 process for a unit time. Process has priority of 4
Executing queue 2 and 3 process for a unit time. Process has priority of 4
Executing queue 2 and 3 process for a unit time. Process has priority of 4
Executing queue 2 and 3 process for a unit time. Process has priority of 4
Executing queue 2 and 3 process for a unit time. Process has priority of 4
Broke from queue 2
Queue 3 completed
Queue 2 completed
Queue 1 in hand
Executing queue 1 and 1 process for a unit time. Process has priority of 7
Broke from queue 1
Queue 2 completed
Queue 1 in hand
Executing queue 1 and 1 process for a unit time. Process has priority of 7
Executing queue 1 and 1 process for a unit time. Process has priority of 7
Executing queue 1 and 1 process for a unit time. Process has priority of 7
Broke from queue 1

Time taken for queue 1 to execute: 28
Process 1 of queue 1 took 28
Process 2 of queue 1 took 6

Time taken for queue 2 to execute: 24
Process 1 of queue 2 took 13
Process 2 of queue 2 took 19
Process 3 of queue 2 took 24

Time taken for queue 3 to execute: 0


Process     | Turn Around Time | Waiting Time
Queue 1
Process P1      28                  20
Process P2      6                   3
Queue 2
Process P1      13                  6
Process P2      19                  13
Process P3      24                  19
Queue 3

 The average turnaround time is : 18

 The average waiting time is : 12